Introduce oauth linking for Google accounts.
This CR does major revamp by performing the following
- Replace svelta-spa-router with svelte-navigator.
- Introduce spa handling on server side introducing web/spa.go

The svelta-spa-router using hash based routing. Hash based
a.k.a fragments are not supported as valid redirect uri in
Google OAuth2 framework. Hence the search for an alternative to
perform oauth2 dance. At one point was also considering using
plain html at the static director location.

Split the web.go to individual components for the respective
handling (api, oauth, spa) which helps maintain the respective
routes in their own files. The refresh will now be sent from
UI instead having this as a startup argument. The clientId and
client secret will remain to be startup argument for security
reasons (client secret should not be in SCM). Having the refresh
token come from UI provides the flexibility of supporting multiple
users and eliminate the manual work required to generate a
refresh token. This also paves way for future to have this as a
hosted service.

Having a hosted service however requires further effort
- local scan as a download tool
- overall authentication for the service e.g. JWT.
- refresh token is currently stored as a client store. This should
be replaced by account token with the refresh token stored on the
server and not leaked to client.